# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000/v1"
    description: "Server locale (v1)"
info:
  title: WASAText API
  description: Struttura API del progetto WASAText del corso di Programmazione Web.
  version: 1.0.0
...
security:
   - LegacyAuth: []
paths:
...
  /users/{userId}/username:
    put:
      tags: ["profile"]
      summary: cambio dell'username
      description: viene modificato il nome dell'utente, se unico, e il vecchio viene liberato per il riuso
      operationId: setMyUserName
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: oggetto contenente il nuovo username 
              type: object
              required: [username]
              properties:
                username:
                  type: string
                  example: Giorgia
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                  description: nuovo username
      responses:
        '200':
          description: username aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                success:
                  value:
                    User:
                      id: 125461561
                      username: nuovo username
                      profilePhoto: "https://example.com/photo.jpg"          
        '409':
          description: username già esistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/photo:
    put:
      tags: ["profile"]
      summary: aggiorna la foto profilo utente
      description: aggiorna la foto profilo utente caricando un file 
      operationId: setMyPhoto
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [photo]
              description: richiesta della foto profilo da caricare
              properties:
                photo:
                  type: string
                  format: url
                  minLength: 1
                  maxLength: 2000
                  pattern: '^.*?$'
                  description: URL foto profilo 
      responses:
        '200':
          description: foto aggiornata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                success:
                  value:
                    User:
                      id: 125461561
                      username: Mario
                      profilePhoto: nuova foto caricata  
...
  /users/{userId}/chats/{chatId}/messages:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/ChatId'
    post:
      tags: ["messages"]
      summary: invia un nuovo messaggio in una conversazione
      description: invia un nuovo messaggio in una conversazione esistente e o la crea se è il primo messaggio
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Oggetto per l'invio di un messaggio
              properties:
                text:
                  type: string
                  description: Contenuto testuale del messaggio
                  minLength: 0
                  maxLength: 2000
                  pattern: "^.*$"
                photoUrl: 
                  type: string
                  description: URL della foto
                  format: url
                  minLength: 0
                  maxLength: 2000
                  pattern: "^.*$"
                replyTo:
                  type: integer
                  description: ID del messaggio a cui si risponde
                isForward:
                  type: boolean
                  description: Indica se il messaggio è inoltrato
      responses:
        '200':
          description: messaggio inviato con successo in chat esistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                success:
                  value:
                    Message:
                      id: 2151531561
                      text: "Hey ciao!"
                      sentBy: Lucia
                      sentAt: 2024-05-20T15:00:00Z
                      checkmark: true
              
        '201':
          description: chat creata e messaggio inviato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                success:
                  value:
                    Message:
                      id: 2151531561
                      text: "Hey ciao!"
                      sentBy: Lucia
                      sentAt: 2024-05-20T15:00:00Z
                      checkmark: true
  /users/{userId}/chats/{chatId}/messages/{messageId}/forward:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/ChatId'
      - $ref: '#/components/parameters/MessageId'
    post:
      tags: ["messages"]
      summary: inoltra un messaggio ad una o più destinazioni
      description: inoltra un messaggio ad una o più destinazioni speicificate con un lista di userId
      operationId: forwardMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targets]
              description: oggetto creante la lista di tutti gli utenti a cui inoltrare il messaggio
              properties:
                targets:
                  type: array
                  minItems: 1
                  maxItems: 2000
                  items:
                    type: integer
                  description: lista di chatId a cui inoltrare il messaggio
      responses:
        '200':
          description: messaggio inoltrato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              examples:
                success:
                  value:
                    Message:
                      id: 2151531561
                      text: "Hey ciao!"
                      sentBy: Lucia
                      sentAt: 2024-05-20T15:00:00Z
                      checkmark: false 
  /users/{userId}/chats/{chatId}/messages/{messageId}/reactions:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/ChatId'
      - $ref: '#/components/parameters/MessageId'
    post:
      tags: ["messages"]
      summary: aggiungi una reazione a un messaggio
      description: aggiungi una reazione a un messaggio selezionando un emoticon sotto forma di unicode o shortcode
      operationId: commentMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emoticon]
              description: oggetto che ritorna l'unicode della emoticon che l'utente desidere usare
              properties:
                emoticon:
                  type: string
                  description: emoticon unicode o shortcode
                  pattern: ".*"
                  minLength: 1
                  maxLength: 64
      responses:
        '201':
          description: reazione aggiunta con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reaction'
              examples:
                success:
                  value:
                    User:
                      reactTo: 2151531561
                      emoticon: 🙂                    

    delete:
      tags: ["messages"]
      summary: rimuove una reazione da un messaggio
      description: rimuove una propria reazione da un messaggio inviato dall'utente corrente
      operationId: uncommentMessage
      responses:
        '204':
          description: reazione rimossa con successo
  /users/{userId}/chats/{chatId}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/ChatId'
      - $ref: '#/components/parameters/MessageId'
    delete:
      tags: ["messages"]
      summary: elimina un messaggio inviato
      description: elimina un proprio messaggio inviato in una delle proprie chat
      operationId: deleteMessage
      responses:
        '204':
          description: messaggio eliminato con successo
...
components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificatore numerico dell'utente
    ChatId:
      name: chatId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificatore numerico della chat
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificatore numerico del messaggio
  schemas:
...
    User:
      type: object
      description: Rappresenta un utente del sistema
      properties:
        id:
          type: integer
          description: Identificatore utente
        username:
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          description: Nome utente unico
        profilePhoto:
          type: string
          format: url
          description: URL verso la foto profilo
        token:
          type: string
          description: Token di sessione per l'autenticazione
...
    Message:
      type: object
      description: Rappresenta un messaggio inviato in una chat
      properties:
        id:
          type: integer
          description: Identificatore del messaggio
        text:
          type: string
          description: Testo del messaggio
          minLength: 0
          maxLength: 2000
        sentAt:
          type: string
          format: date-time
          description: Data/ora di invio
        sentBy:
          type: integer
          description: id utente che ha inviato
        checkmark:
          type: boolean
          description: stato di consegna/lettura
        senderName:
          type: string
          description: Nome visuale del mittente (per i gruppi)
        replyTo:
          type: integer
          description: ID del messaggio a cui questo risponde
        isForward:
          type: boolean
          description: True se il messaggio è inoltrato
        photoUrl:
          type: string
          description: URL della foto allegata
    Reaction:
      type: object
      description: Rappresenta una reazione a un messaggio
      properties:
        reactTo:
          type: integer
          description: id del messaggio a cui reagire
        emoticon:
          type: string
          description: emoticon unicode o shortcode
          minLength: 1
          maxLength: 64
    Error:
      type: object
      description: schema di errore generico
      properties:
        code:
          type: integer
          description: Codice di errore
        message:
          type: string
          minLength: 5
          maxLength: 200
          pattern: "^.*$"
          description: messaggio di spiegazione dell'errore
  securitySchemes: 

    LegacyAuth:
      type: apiKey
      name: sessionToken
      in: header
      description: valore del token inviato dal server nel momento di login